requirements/common.txt
# Primary packages
gspread==3.1.0
oauth2client==4.1.3
pylama==7.7.1
pyzmq==18.1.1

# Secondary packages for pylama
mccabe==0.6.1
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
snowballstemmer==2.0.0

# Secondary packages for gspread
certifi==2019.9.11
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.7

# Secondary packages for oauth2client
httplib2==0.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
rsa==4.0
six==1.13.0
